+++ Dfe.Academies.Academisation.Data/Migrations/20260722153030_AddSignificantChangeSchoolName.cs
@@ -1,0 +1,31 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace Dfe.Academies.Academisation.Data.Migrations
+{
+    /// <inheritdoc />
+    public partial class AddSignificantChangeSchoolName : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.AddColumn<string>(
+                name: "SchoolName",
+                schema: "academisation",
+                table: "SignificantChangeProject",
+                type: "nvarchar(max)",
+                nullable: false,
+                defaultValue: "");
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropColumn(
+                name: "SchoolName",
+                schema: "academisation",
+                table: "SignificantChangeProject");
+        }
+    }
+}