+++ Dfe.Academies.Academisation.Data/Repositories/SignificantChangeProjectRepository.cs
@@ -1,5 +1,6 @@
 ﻿using Dfe.Academies.Academisation.Domain.SeedWork;
 using Dfe.Academies.Academisation.Domain.SignificantChange;
+using Microsoft.EntityFrameworkCore;
 
 namespace Dfe.Academies.Academisation.Data.Repositories
 {
@@ -8,5 +9,24 @@ namespace Dfe.Academies.Academisation.Data.Repositories
 		private readonly AcademisationContext _context = context ?? throw new ArgumentNullException(nameof(context));
 
 		public IUnitOfWork UnitOfWork => _context;
+
+		public async Task<(IEnumerable<SignificantChangeProject> projects, int totalCount)> SearchSignificantChangeProjects(int page, int count, CancellationToken cancellationToken)
+		{
+			IQueryable<SignificantChangeProject> queryable = dbSet;
+
+			int totalProjects = await queryable.CountAsync(cancellationToken);
+			var projects = await queryable
+				.OrderByDescending(acp => acp.CreatedOn)
+				.Skip((page - 1) * count)
+				.Take(count)
+				.ToListAsync(cancellationToken);
+
+			return (projects, totalProjects);
+		}
+
+		public async Task<SignificantChangeProject?> GetSignificantChangeProjectById(int id, CancellationToken cancellationToken)
+		{
+			return await dbSet.SingleOrDefaultAsync(project => project.Id == id, cancellationToken);
+		}
 	}
 }