+++ Dfe.Academies.Academisation.Domain/SignificantChange/ISignificantChangeProjectRepository.cs
@@ -4,5 +4,7 @@ namespace Dfe.Academies.Academisation.Domain.SignificantChange
 {
 	public interface ISignificantChangeProjectRepository : IRepository<SignificantChangeProject>, IGenericRepository<SignificantChangeProject>
 	{
+		Task<(IEnumerable<SignificantChangeProject> projects, int totalCount)> SearchSignificantChangeProjects(int page, int count, CancellationToken cancellationToken);
+		Task<SignificantChangeProject?> GetSignificantChangeProjectById(int id, CancellationToken cancellationToken);
 	}
 }