+++ Dfe.Academies.Academisation.WebApi/Controllers/SignificantChangeController.cs
@@ -1,8 +1,10 @@
-using Dfe.Academies.Academisation.Core;
+﻿using Dfe.Academies.Academisation.Core;
 using Dfe.Academies.Academisation.IService.ServiceModels.SignificantChange;
+using Dfe.Academies.Academisation.Service.Queries.SignificantChange;
 using Dfe.Academies.Academisation.Service.Commands.SignificantChange;
 using MediatR;
 using Microsoft.AspNetCore.Mvc;
+using Dfe.Academies.Academisation.IService.ServiceModels.Legacy.ProjectAggregate;
 
 namespace Dfe.Academies.Academisation.WebApi.Controllers
 {
@@ -36,6 +38,33 @@ namespace Dfe.Academies.Academisation.WebApi.Controllers
 				null => BadRequest(),
 				_ => throw new NotImplementedException()
 			};
+		}
+
+		[HttpPost("search", Name = "GetSignificantChangeProjects")]
+		[ProducesResponseType(StatusCodes.Status200OK)]
+		public async Task<ActionResult<PagedDataResponse<SignificantChangeProjectSearchResponse>>> GetSignificantChangeProjects(
+			[FromBody] GetSignificantProjectsQuery query,
+			CancellationToken cancellationToken)
+		{
+			_logger.LogInformation("Getting significant change projects");
+			var result = await _mediator.Send(query, cancellationToken);
+
+			return result switch
+			{
+				PagedDataResponse<SignificantChangeProjectSearchResponse> pagedResult => Ok(pagedResult)
+			};
+		}
+
+		[HttpGet("{id:int}", Name = "GetSignificantChangeProjectById")]
+		[ProducesResponseType(StatusCodes.Status200OK)]
+		[ProducesResponseType(StatusCodes.Status404NotFound)]
+		public async Task<ActionResult<SignificantChangeProjectSearchResponse>> GetSignificantChangeProject(int id,
+			CancellationToken cancellationToken)
+		{
+			_logger.LogInformation("Getting significant change project with id {Id}", id);
+			var result = await _mediator.Send(new GetSignificantChangeProjectByIdQuery(id), cancellationToken);
+
+			return result is null ? NotFound() : Ok(result);
 		}
 	}
 }